oradb-manage-oraclenet: new role for managing oraclenet files
The includes for listener.ora, sqlnet.ora and tnsnames.ora has been
moved from role oradb-manage-db to oradb-maange-oraclenet with
include_role in oradb-manage-db.
This change is compatible to all existing playbooks for ansible-oracle.

It is the 1st step for implementing full oracle Restart and RAC support
for managing the listener.ora and easier configuration of oraclenet
files after the initial installation.
The role oradb-manage-oraclenet could be used to configure the oraclenet
files only instead of using oradb-manage-db with tags for filtering the
tasks.

Some variables has been moved from defaults to
orasw-meta/defaults/main.yml. This code cleanup is a preparation for
later cleanup of other roles.