ansible-oracle: BREAKING CHANGE! move to a collection
This is one of the biggest changes during the life of ansible-oracle.
The project moves fto a collection, which leads to the following breaking changes:

- rename all roles with '-' in name to '_'
This is a mandatory change, when moving to a collection.
galaxy  will not import a collection with a '-' in rolename...
- moved playbooks to playbooks folder
This is the 1st stept to reorganize the old - mostly not working -
playbooks. The redesign of playbooks and example inventory is planned
for a later time.